Add tagging and improve version capture
This patch loosens the expression in checkVersionNumber to only look for
'Koha::Plugins::Base'.  Often a perltidy would result in additional
whitespace being introduced into that line and looking for just the
module base should be enough.

We also escapte the special bracket characters to prevent an error on
echo and update the git log to look for all commits (not just the most
recent one) and grep to stop at the first case of finding a version
number commit (We expand from 4 digits to 6 for x.xx.xx type versioning
too)

Finally, we add a git tag command to tag the release as well as
commiting it.