fix: Only run if new version > old. regex fixes for ripgrep users.
This patch will ensure that the regex portions run regardless of if the
user is running ripgrep or not.
This also adds checking that the new version is a higher version number
than the old number. Also nicer login for if this is the first run and
there have been no versions committed yet.